requirements.txt
matplotlib==3.10.3
numpy==2.3.1
opencv_contrib_python==4.10.0.84
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
paddleocr==3.0.2
Pillow==11.2.1
streamlit==1.46.0
ultralytics==8.3.159
